refactoring CircularCloudLayouter
